Pulling values from other cells.
This commit changes, when and how dependency callbacks are called.
With this implementation, cells stage new values in one-element-queues and inform dependent cells to eventually dequeue this value and pass it to the callback. Cells do not call callbacks after they have been completed, as has been decided by phaller. This makes Updater.ignoreIfFinal obsolete.